+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [3.6.0] - 2021-09-07
+- Updated to work in Mongo 4.2.14 environment.
+- Updated to work in a SSL environment.
+- Added ability to override the default mail command and use mailx.
+
 ### Changed
+- config/mongo.py.TEMPLATE:  Added SSL connection entries.
 - Removed unnesscary \*\*kwargs from argument lists.
+- Documentation updates.
 
 
 ## [3.5.0] - 2021-01-08
@@ -55,7 +61,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Changed
 - run_program:  Added repset_hosts to mongo_class.RepSet instance call.
-- chk_mem_rep_lag:  Changed dictionary to CamelCase.
+- chk_mem_rep_lag:  Changed dictionary to PascalCase.
 - chk_mem_rep_lag:  Refactored to handle only the JSON format and convert to standard out format if requested.
 - Documentation updates.